fixed bug in FM where alerts_dir was not already created - added start services for boiler-alerts and boiler-web
There was an issue in feedManagement that was unaccounted for, where if the alerts_dir was not created by an alert being stored (i.e. if there were no alerts the dir would not be made), it would except, as it was unable to list the dir since it didn't exist.

Forgot to start the systemd services in the install.sh script. Should be working now.