fixed a linux quirk because rm/rmdir 'is a directory' and 'directory not empty'.
Yeah because apparently this is a thing. I remember when I first used Linux, I tried to delete a folder with "rm" and it said 'is a directory'. Well, duh, it's a directory! I want to delete it dumbass. Then I typed "rmdir" and then it said "directory not empty" and I was like, bro! Why are you stupid? So then I googled it and found out I have to do "rm -rf". And now I understand why "sudo rm -rf /" is laughable.